+++ src/Repositories/CommonTrait.php
@@ -8,14 +8,15 @@ trait CommonTrait
 {
     protected int $maxttl = 3600 * 24 * 30;
 
-    public function setMaxTtl(\DateInterval|int $ttl)
+    public function setMaxTtl(\DateInterval|int $ttl): self
     {
         if (empty($ttl)) {
             $ttl = 0;
         } elseif ($ttl instanceof \DateInterval) {
             $ttl = (int) $ttl->format("%s");
         }
-        $this->maxTtl = $ttl;
+        $this->maxttl = $ttl;
+        return $this;
     }
     protected function maxTtl(\DateInterval|null|int $ttl = null): int
     {